Exercise 2.2

+++ Beginners/Day-2/Coding Exercise 2.2 BMI Calculator.py
@@ -4,5 +4,5 @@ weight = input("enter your weight in kg: ")
 # 🚨 Don't change the code above 👆
 
 # Write your code below this line 👇
-BMI = int(weight) / float(height) ** 2
+BMI = float(weight) / float(height) ** 2
 print(int(BMI))
